% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gev.R
\name{gev}
\alias{gev}
\alias{gevLoglik}
\alias{dGEV}
\alias{pGEV}
\title{Generalised Extreme Value functions}
\usage{
gevLoglik(x, loc, scale, shape, individual = TRUE, ...)

dGEV(x, loc = 0, scale = 1, shape = 0, log = FALSE, ...)

pGEV(q, loc = 0, scale = 1, shape = 0, lower.tail = TRUE, log.p = FALSE, ...)
}
\arguments{
\item{x}{A numeric vector of observations. Typically, these are
block maxima, that is, the largest observation in a block of contiguous
observations.}

\item{loc}{A numeric vector. Values of the GEV location parameter \eqn{\mu}.}

\item{scale}{A numeric vector. Values of the GEV scale parameter
\eqn{\sigma}.}

\item{shape}{A numeric vector. Values of the GEV shape parameter \eqn{\xi}.}

\item{individual}{A logical scalar. Relevant to \code{gevLoglik} and
\code{gevScore}. If \code{individual = FALSE} then only the sum of
contributions from all observations in \code{x} is calculated.  If
\code{individual = TRUE} then individual contributions from each
observation in \code{x} are calculated.}

\item{...}{Further arguments to be passed to \code{\link{log1pdx}}, which evaluates
terms of the form \eqn{\log(1+x)/x} in the GEV log-likelihood.}
}
\value{
\strong{Log-likelihood} (\code{gevLoglik}). If
\code{individual = FALSE} the value of the log-likelihood. If
\code{individual = TRUE} a vector of length \verb{length\{x\}}
containing the contributions to the log-likelihood from each of the
observations.

\strong{Score} (\code{gevScore}).  If \code{individual = FALSE} the value
of the score, a vector of length 2 containing the derivative of the
log-likelihood evaluated at the input parameter values.
If \code{individual = TRUE} the values of the contributions to the score
from each of the observations, a
\code{length(x)}\eqn{ \times 2}{ x 2} matrix.
The columns are named \code{sigma[u]} and \code{xi}.

\strong{Observed information} (\code{gevInfo}).  The observed information: a
\eqn{2 \times 2}{2 x 2} matrix with row and column names
\code{c(sigma[u], xi)}.
}
\description{
Calculate the log-likelihood function, score and observed information for
a random sample from a generalised extreme value (GEV) distribution,
including cases where the shape parameter is very close to zero.
}
\details{
\strong{GEV density} (\code{dGEV}) and \strong{GEV distribution function} (\code{pGEV}).
The input vectors \code{x}, \code{loc}, \code{scale} and \code{shape} are recycled, so that
the length of the returned vector is the maximum of the lengths of these
arguments. For any element of \code{scale} that is non-positive, \code{NaN} is
returned, with no warning.

\strong{Log-likelihood} (\code{gevLoglik}). The two problematic
terms of the log-likelihood both involve \eqn{\log(1+z)/z},
where \eqn{z = \xi (y - \mu)/ \sigma} and where \eqn{y} is a
sample maximum. In one part this is exponentiated, in the other it is not.
The function \code{\link{log1pdx}} is used to calculate \eqn{\log(1+z)/z}, with
special case taken for cases where \code{z} is close to zero.
}
\examples{
### Simulate some data

set.seed(17042022)
y <- rGenExtremeValue(100, 0, 1, 0)

### Log-likelihood

# Calculation using log1pdx()
gevLoglik(y, 0, 1, 1e-8, )
gevLoglik(y, 0, 1, -1e-8)
gevLoglik(y, 0, 1, 0)

# Direct calculation, involving (1 / xi) * log1p(xi * (y - mu) / sigma)
# Mostly fine, but breaks down eventually
gevLoglikDirect(pars = c(0, 1, 1e-309), maxima = y)
gevLoglikDirect(pars = c(0, 1, -1e-309), maxima = y)
}
