% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gevDerivatives.R
\name{gevDerivatives}
\alias{gevDerivatives}
\alias{gev1}
\alias{gev2}
\alias{gev3}
\alias{gev11}
\alias{gev12}
\alias{gev22}
\alias{gev13}
\alias{gev23}
\alias{gev33}
\title{Generalised Extreme Value log-likelihood derivatives}
\usage{
gev1(x, loc = 0, scale = 1, shape = 0, ...)

gev2(x, loc = 0, scale = 1, shape = 0, ...)

gev3(x, loc = 0, scale = 1, shape = 0, ...)

gev11(x, loc = 0, scale = 1, shape = 0, ...)

gev12(x, loc = 0, scale = 1, shape = 0, ...)

gev22(x, loc = 0, scale = 1, shape = 0, ...)

gev13(x, loc = 0, scale = 1, shape = 0, ...)

gev23(x, loc = 0, scale = 1, shape = 0, ...)

gev33(x, loc = 0, scale = 1, shape = 0, ...)
}
\arguments{
\item{x}{A numeric vector of observations. Typically, these are
block maxima, that is, the largest observation in a block of contiguous
observations.}

\item{loc}{A numeric vector. Value(s) of the GEV location parameter \eqn{\mu}.}

\item{scale}{A numeric vector of \strong{positive} values. Value(s) of the
GEV scale parameter \eqn{\sigma}.}

\item{shape}{A numeric vector. Value(s) of the GEV shape parameter
\eqn{\xi}.}

\item{...}{Further arguments to be passed to \code{\link{log1pdx}}, \code{\link{dlog1pdx}} or
\code{\link{d2log1pdx}}. See \strong{Details}.}
}
\value{
A numeric vector of length \code{length(x)} containing the relevant first or
second derivative of the GEV log-density.
}
\description{
Calculates the first and second derviatives of the log-density of a
generalised extreme value (GEV) distribution, including cases where the
shape parameter is very close to zero.
}
\details{
The input vectors \code{x}, \code{loc}, \code{scale} and \code{shape} are not inside these
functions to force them to have the same lengths. Therefore, each of these
input vectors should either have the same length as the longest of these
vectors, or have length 1. These vectors should also satisfy the parameter
constraints \eqn{\sigma > 0} and \eqn{1 + \xi(x - \mu) / \sigma > 0}.
These requirements are explicitly not checked inside these functions,
so the user must take care when calling them.

\strong{First derivatives}. The functions \code{gev1}, \code{gev2} and \code{gev3} are used in
\code{\link{gevScore}} to calculate the first derivatives of the GEV log-likelihood
\eqn{\ell} with respect to \eqn{\mu}, \eqn{\sigma} and \eqn{\xi}. \code{gevi}
provides \eqn{\partial \ell / \partial \theta_i}, where
\eqn{\theta = (\mu, \sigma, \xi)}.

The first derivatives of the log-likelihood with respect to \eqn{\mu} and
\eqn{\sigma} both involve \eqn{\log(1+z)/z}, calculated using \code{\link{log1pdx}}.
The first derivative of the log-likelihood with respect to \eqn{\xi}
involves \eqn{\log(1+z)/z} and also
\eqn{\log(1+z)/z^2 - z^{-1}(1+z)^{-1}}. The latter is calculated using
\code{\link{dlog1pdx}}.

\strong{Second derivatives}. The functions \code{gev11}, \code{gev12}, \code{gev13}, \code{gev22},
\code{gev23} and \code{gev33} are used in \code{\link{gevObsInfo}} to calculate the second
derivatives of the GEV log-likelihood \eqn{\ell} with respect to
\eqn{\mu}, \eqn{\sigma} and \eqn{\xi}. \code{gevij} provides
\eqn{\partial^2 \ell / \partial \theta_i \theta_j}, where
\eqn{\theta = (\mu, \sigma, \xi)}.

All the second derivatives of the log-likelihood involve
\eqn{\log(1+z)/z}, calculated using \code{\link{log1pdx}}. Any second derivative
made with respect to \eqn{\xi} also involves
\eqn{\log(1+z)/z^2 - z^{-1}(1+z)^{-1}}, which is calculated using
\code{\link{dlog1pdx}}. Differentiating the log-likelihood twice with respect to
\eqn{\xi} produces the function
\eqn{2 \log(1+z)/z^3 - 2 z^{-2}(1+z)^{-1} - z^{-1}(1+z)^{-2}}, which is
calculated using \code{\link{d2log1pdx}}.
}
\examples{
### Simulate some data
set.seed(28122023)
x <- rGEV(10)

# First derivatives
gev1(x)
gev2(x)
gev3(x)

# Second derivatives
gev11(x)
gev12(x)
gev13(x)
gev22(x)
gev23(x)
gev33(x)
}
