% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalisedExtremeValueExpectedInformation.R
\name{gevExpectedInformation}
\alias{gevExpectedInformation}
\alias{imm}
\alias{iss}
\alias{ixx}
\alias{ims}
\alias{imx}
\alias{isx}
\title{Generalised Extreme Value Expected Information}
\usage{
imm(sigma, xi)

iss(sigma, xi, eps = 0.001)

ixx(sigma, xi, eps = 0.001)

ims(sigma, xi, eps = 0.001)

imx(sigma, xi, eps = 0.001)

isx(sigma, xi, eps = 0.001)
}
\description{
Describe
}
