% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalisedExtremeValueExpectedInformation.R
\name{gevExpInfo}
\alias{gevExpInfo}
\alias{imm}
\alias{iss}
\alias{ixx}
\alias{ims}
\alias{imx}
\alias{isx}
\title{GEV Distribution Expected Information}
\usage{
imm(sigma, xi)

iss(sigma, xi, eps = 0.003)

ixx(sigma, xi, eps = 0.003)

ims(sigma, xi, eps = 0.003)

imx(sigma, xi, eps = 0.003)

isx(sigma, xi, eps = 0.003)

gevExpInfo(sigma, xi, eps = 0.003)
}
\arguments{
\item{sigma, xi}{Numeric vectors.
\itemize{
\item \code{gevExpInfo}: \code{sigma} and \code{xi} must have length 1.
\item \code{imm}, \code{iss}, \code{ixx}, \code{ims}, \code{imx}, \code{isx}:
}}

\item{eps}{tolerance}
}
\description{
Calculates the expected information matrix for the GEV distribution.
}
