% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gevLikelihood.R
\name{gevLikelihood}
\alias{gevLikelihood}
\alias{gevLoglik}
\alias{gevScore}
\title{Generalised Extreme Value functions}
\usage{
gevLoglik(x, loc = 0, scale = 1, shape = 0, sum = FALSE, ...)

gevScore(x, loc = 0, scale = 1, shape = 0, sum = FALSE, ...)
}
\arguments{
\item{x}{A numeric vector of observations. Typically, these are
block maxima, that is, the largest observation in a block of contiguous
observations.}

\item{loc}{A numeric vector. Values of the GEV location parameter \eqn{\mu}.}

\item{scale}{A numeric vector of \strong{positive} values. Values of the
GEV scale parameter \eqn{\sigma}.}

\item{shape}{A numeric vector. Values of the GEV shape parameter \eqn{\xi}.}

\item{sum}{A logical scalar. Relevant to \code{gevLoglik} and
\code{gevScore}. If \code{sum = TRUE} then only the sum of
contributions from all observations in \code{x} is calculated.  If
\code{sum = FALSE} then individual contributions from each
observation in \code{x} are calculated.}

\item{...}{Further arguments to be passed to \code{\link{log1pdx}}, which evaluates
terms of the form \eqn{\log(1+x)/x} in the GEV log-likelihood.}
}
\value{
\strong{Log-likelihood} (\code{gevLoglik}). If
\code{sum = TRUE} a scalar, the value of the log-likelihood. If
\code{sum = FALSE} a vector of length \verb{length\{x\}}
containing the contributions to the log-likelihood from each of the
observations.

\strong{Score} (\code{gevScore}).  If \code{sum = TRUE} the value
of the score, a vector of length 3 containing the derivative of the
log-likelihood evaluated at the input parameter values.
If \code{sum = FALSE} the values of the contributions to the score
from each of the observations, a
\code{length(x)}\eqn{ \times 3} matrix.
The columns are named \code{sigma[u]} and \code{xi}.

\strong{Observed information} (\code{gevInfo}).  The observed information: a
\eqn{2 \times 2} matrix with row and column names
\code{c(sigma[u], xi)}.
}
\description{
Calculate the log-likelihood, score and observed information for
a random sample from a generalised extreme value (GEV) distribution,
including cases where the shape parameter is very close to zero.
}
\details{
\strong{Log-likelihood} (\code{gevLoglik}). The two problematic
terms of the log-likelihood both involve \eqn{\log(1+z)/z},
where \eqn{z = \xi (x - \mu)/ \sigma}. In one part this is exponentiated,
in the other it is not. The function \code{\link{log1pdx}} is used to calculate
\eqn{\log(1+z)/z}, with special case taken for cases where \code{z} is close to
zero.

\strong{Score} (\code{gevScore}). The first derivatives of the log-likelihood with
respect to \eqn{\mu} and \eqn{\sigma} both involve \eqn{\log(1+z)/z},
calculated using \code{\link{log1pdx}}. The first derivative of the log-likelihood
with respect to \eqn{\xi} involves \eqn{\log(1+z)/z} and also
\eqn{\log(1+z)/z^2 - z^{-1}(1+z)^{-1}}. The latter is calculated using
\code{\link{log1pdx2}}.
}
\examples{
### Simulate some data

set.seed(17042022)
y <- rGenExtremeValue(100, 0, 1, 0)

### Log-likelihood

# Calculation using log1pdx()
gevLoglik(y, 0, 1, 1e-8, )
gevLoglik(y, 0, 1, -1e-8)
gevLoglik(y, 0, 1, 0)

# Direct calculation, involving (1 / xi) * log1p(xi * (y - mu) / sigma)
# Mostly fine, but breaks down eventually
gevLoglikDirect(pars = c(0, 1, 1e-309), maxima = y)
gevLoglikDirect(pars = c(0, 1, -1e-309), maxima = y)

# Score
set.seed(28122023)
x <- rGEV(10)
gevScore(x)
gevScore(x, sum = TRUE)
}
