% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpLikelihood.R
\name{gpLikelihood}
\alias{gpLikelihood}
\alias{gpLoglik}
\alias{gpScore}
\alias{gpObsInfo}
\title{Generalised Pareto functions}
\usage{
gpLoglik(x, scale = 1, shape = 0, sum = FALSE, ...)

gpScore(x, scale = 1, shape = 0, sum = FALSE, ...)

gpObsInfo(x, scale = 1, shape = 0, sum = FALSE, ...)
}
\arguments{
\item{x}{A numeric vector of observations. Typically, these are
block maxima, that is, the largest observation in a block of contiguous
observations.}

\item{scale}{A numeric vector of \strong{positive} values. Values of the
GP scale parameter \eqn{\sigma}.}

\item{shape}{A numeric vector. Values of the GP shape parameter \eqn{\xi}.}

\item{sum}{A logical scalar. Relevant to \code{gpLoglik} and \code{gpScore}. If
\code{sum = TRUE} then only the sum of contributions from all combinations of
observations in \code{x } and parameter values in \code{loc}, \code{scale} and \code{shape}
is calculated.  If \code{sum = FALSE} then these individual contributions are
summed to produce the result.}

\item{...}{Further arguments to be passed to \code{\link{log1pdx}}, \code{\link{dlog1pdx}} or
\code{\link{d2log1pdx}}. See \code{\link{gpDerivatives}}.}
}
\value{
The number \code{n}, say, of combinations of the input parameters \code{x},
\code{scale} and \code{shape} in the returned object is the maximum of the lengths
of these arguments, with these arguments being recycled as necessary.

\strong{Log-likelihood} (\code{gpLoglik}). If \code{sum = FALSE} a vector of length \code{n}
containing individual log-likelihood contributions. If \code{sum = TRUE} a
scalar, the value of sum of these contributions.

\strong{Score} (\code{gpScore}). If \code{sum = FALSE} the values of the contributions
to the score, an \code{n}\eqn{ \times 2} matrix. The columns are named
\code{loc}, \code{scale} and \code{shape}. If \code{sum = TRUE}, a vector of length 2
containing the column sums of the matrix returned if \code{sum = FALSE}.

\strong{Observed information} (\code{gpObsInfo}).  If \code{sum = FALSE} the values of
the contributions to the observed information, an
\code{n} \eqn{ \times 2 \times 2} array. The second and third dimensions are
named columns are named \code{scale} and \code{shape}. If \code{sum = TRUE}, a
\eqn{2 \times 2} matrix giving the observed information matrix, obtained
by applying \code{\link{colSums}} to the array returned if \code{sum = FALSE}.

\code{NA} is returned for any component of the inputs for which \code{x}, \code{loc},
\code{scale} or \code{shape} is \code{NA}.

\code{NaN} is returned for any component of \code{scale} that is non-positive and
for any component for which \code{loc}, \code{scale} and/or \code{shape} are either \code{Inf}
or \code{-Inf}, with no warning given.
}
\description{
Calculate the log-likelihood, score and observed information for
a random sample from a generalised Pareto (GP) distribution,
including cases where the shape parameter is very close to zero.
}
\details{
\strong{Log-likelihood} (\code{gpLoglik}). The two problematic
terms of the log-likelihood both involve \eqn{\log(1+z)/z},
where \eqn{z = \xi x / \sigma}. In one part this is exponentiated,
in the other it is not. The function \code{\link{log1pdx}} is used to calculate
\eqn{\log(1+z)/z}, with special case taken for cases where \code{z} is close to
zero.

\strong{Score} (\code{gpScore}). The first derivative of the log-likelihood
with respect to \eqn{\xi} involves \eqn{\log(1+z)/z}, calculated using
\code{\link{log1pdx}}.

The functions \code{gp1} and \code{gp2} are used in
\code{gpScore} to calculate the first derivatives of the GP log-likelihood
\eqn{\ell} with respect to \eqn{\sigma} and \eqn{\xi}.
See \code{\link{gpDerivatives}}.

\strong{Observed information} (\code{gpObsInfo}). The second derivative of the
log-likelihood respect to \eqn{\xi} involves the function
\eqn{2 \log(1+z)/z^3 - 2 z^{-2}(1+z)^{-1} - z^{-1}(1+z)^{-2}}, which is
calculated using \code{\link{d2log1pdx}}.

The functions \code{gp11}, \code{gp12} and \code{gp22} are
used in \code{gpObsInfo} to calculate the second derivatives of the GP
log-likelihood \eqn{\ell} with respect to \eqn{\sigma} and
\eqn{\xi}. \code{gpij} provides
\eqn{\partial^2 \ell / \partial \theta_i \theta_j}, where
\eqn{\theta = (\sigma, \xi)}.
See \code{\link{gpDerivatives}}.
}
\examples{
### Simulate some data
set.seed(28122023)
x <- rGP(5)

### Log-likelihood

# Calculation using log1pdx()
gpLoglik(x, 1, 1e-8)
gpLoglik(x, 1, -1e-8)
gpLoglik(x, 1, 0)

# Direct calculation, involving (1 / xi) * log1p(xi * (x - mu) / sigma)
# Mostly fine, but breaks down eventually
gpLoglikDirect(pars = c(1, 1e-309), excesses = x)
gpLoglikDirect(pars = c(1, -1e-309), excesses = x)

# Score

gpScore(x)
gpScore(x, sum = TRUE)

# Information

gpObsInfo(x)
gpObsInfo(x, sum = TRUE)
}
