% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evils-internal.R
\name{evils-internal}
\alias{evils-internal}
\alias{log1pxOverx}
\alias{log1pxOverxApprox}
\alias{log1pxOverxLogFunction}
\alias{log1pxOverxDirect}
\title{Internal evils functions}
\usage{
log1pxOverx(x, tol = 1e-04, epsilon = 1e-15)

log1pxOverxApprox(x, epsilon = 1e-15)

log1pxOverxLogFunction(n, x)

log1pxOverxDirect(x)
}
\description{
Internal evils functions
}
\details{
These functions are not intended to be called by the user.
}
\keyword{internal}
