% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evils-internal.R
\name{evils-internal}
\alias{evils-internal}
\alias{log1pxOverx}
\alias{log1pxOverxApprox}
\alias{log1pxOverxLogFunction}
\alias{log1pxOverxDirect}
\alias{gpLoglikDirect}
\alias{gevLoglikDirect}
\alias{nhppLoglikDirect}
\alias{rGenPareto}
\alias{qGenPareto}
\alias{BoxCoxVec}
\alias{rGenExtremeValue}
\alias{qGenExtremeValue}
\title{Internal evils functions}
\usage{
log1pxOverx(x, tol = 1e-04, epsilon = 1e-15)

log1pxOverxApprox(x, epsilon = 1e-15)

log1pxOverxLogFunction(n, x)

log1pxOverxDirect(x)

gpLoglikDirect(pars, excesses, individual = FALSE)

gevLoglikDirect(pars, maxima, individual = FALSE)

nhppLoglikDirect(pars, data, u, b = 1, nb = 365, individual = FALSE)

rGenPareto(n, loc = 0, scale = 1, shape = 0)

qGenPareto(p, loc = 0, scale = 1, shape = 0, lower.tail = TRUE, log.p = FALSE)

BoxCoxVec(x, lambda = 1, lambda_tol = 1e-06)

rGenExtremeValue(n, loc = 0, scale = 1, shape = 0)

qGenExtremeValue(
  p,
  loc = 0,
  scale = 1,
  shape = 0,
  lower.tail = TRUE,
  log.p = FALSE
)
}
\description{
Internal evils functions
}
\details{
These functions are not intended to be called by the user.
}
\keyword{internal}
