% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evils-internal.R
\docType{data}
\name{evils-internal}
\alias{evils-internal}
\alias{EulersConstant}
\alias{AperysConstant}
\alias{pxi}
\alias{qxi}
\alias{iss0Fn}
\alias{iss0Constant}
\alias{issFn}
\alias{ixx0Fn}
\alias{ixx0Constant}
\alias{ixxFn}
\alias{ims0Fn}
\alias{ims0Constant}
\alias{imsFn}
\alias{imx0Fn}
\alias{imx0Constant}
\alias{imxFn}
\alias{isx0Fn}
\alias{isx0Constant}
\alias{isxFn}
\alias{gevExpInfoComp}
\alias{lagrangianInterpolation}
\alias{gpLoglikDirect}
\alias{gevLoglikDirect}
\alias{nhppLoglikDirect}
\alias{rGenPareto}
\alias{qGenPareto}
\alias{BC}
\alias{BoxCoxVec}
\alias{rGenExtremeValue}
\alias{qGenExtremeValue}
\title{Internal evils functions}
\format{
An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.
}
\usage{
EulersConstant

AperysConstant

pxi(xi)

qxi(xi)

iss0Fn()

iss0Constant

issFn(xi)

ixx0Fn()

ixx0Constant

ixxFn(xi)

ims0Fn()

ims0Constant

imsFn(xi)

imx0Fn()

imx0Constant

imxFn(xi)

isx0Fn()

isx0Constant

isxFn(xi)

gevExpInfoComp(fun, fun0, xi, eps = 0.003)

lagrangianInterpolation(x0, y0)

gpLoglikDirect(pars, excesses, individual = FALSE)

gevLoglikDirect(pars, maxima, individual = FALSE)

nhppLoglikDirect(pars, data, u, b = 1, nb = 365, individual = FALSE)

rGenPareto(n, loc = 0, scale = 1, shape = 0)

qGenPareto(p, loc = 0, scale = 1, shape = 0, lower.tail = TRUE, log.p = FALSE)

BC(x, lambda, eps = 1e-06)

BoxCoxVec(x, lambda = 1, lambda_tol = 1e-06)

rGenExtremeValue(n, loc = 0, scale = 1, shape = 0)

qGenExtremeValue(
  p,
  loc = 0,
  scale = 1,
  shape = 0,
  lower.tail = TRUE,
  log.p = FALSE
)
}
\description{
Internal evils functions
}
\details{
These functions are not intended to be called by the user.
}
\keyword{internal}
