% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gevExpInfo.R
\name{gevExpInfo}
\alias{gevExpInfo}
\alias{gev11e}
\alias{gev22e}
\alias{gev33e}
\alias{gev12e}
\alias{gev13e}
\alias{gev23e}
\title{GEV Distribution Expected Information}
\usage{
gev11e(sigma, xi)

gev22e(sigma, xi, eps = 0.003)

gev33e(xi, eps = 0.003)

gev12e(sigma, xi, eps = 0.003)

gev13e(sigma, xi, eps = 0.003)

gev23e(sigma, xi, eps = 0.003)

gevExpInfo(sigma, xi, eps = 0.003)
}
\arguments{
\item{sigma, xi}{Numeric vectors. Respective values of the GEV parameters
\eqn{\sigma} and \eqn{\xi}. For \code{gevExpInfo}, \code{sigma} and \code{xi} must have
length 1.}

\item{eps}{A numeric scalar. For values of \eqn{\xi} in \code{xi} that lie in
\verb{(-eps, eps)} an approximation is used instead of a direct calculation.
See \strong{Details}. If \code{eps} is a vector then only the first element is used.}
}
\value{
\code{gevExpInfo} returns a 3 by 3 numeric matrix with row and column
names \verb{mu, sigma, xi}. The other functions return a numeric vector of
length equal to the maximum of the lengths of the arguments, excluding
\code{eps}.
}
\description{
Calculates the expected information matrix for the GEV distribution.
}
\details{
\code{gevExpInfo} calculates, for single pair of values
\eqn{(\sigma, \xi) = } \verb{(sigma, xi)}, the expected information matrix for a
single observation from a GEV distribution with distribution function
\deqn{F(x) = P(X \leq x) = \exp\left\{ -\left[ 1+\xi\left(\frac{x-\mu}{\sigma}\right)
  \right]_+^{-1/\xi} \right\},}
where \eqn{x_+ = \max(x, 0)}.
The GEV expected information is defined only for \eqn{\xi > -0.5} and does
not depend on the value of \eqn{\mu}.

The other functions are vectorised and calculate the individual
contributions to the expected information matrix. For example, \code{imm}
calculates the expectation \eqn{i_{\mu\mu}} of the negated second
derivative of the GEV log-density with respect to \eqn{\mu}, that is, each
\code{m} indicates one derivative with respect to \eqn{\mu}. Similarly, \code{s}
denotes one derivative with respect to \eqn{\sigma} and \code{x} one derivative
with respect to \eqn{\xi}, so that, for example, \code{isx} calculates the
expectation \eqn{i_{\sigma\xi}} of the negated GEV log-density after one
taking one derivative with respect to \eqn{\sigma} and one derivative with
respect to \eqn{\xi}. Note that \eqn{i_{\xi\xi}} depends only on
\eqn{\xi}.

The expectation in \code{imm} can be calculated in a directly way for all
\eqn{\xi > -0.5}. For the other components, direct calculation of the
expectation is unstable when \eqn{\xi} is close to 0. Instead, we use
a quadratic approximation over \verb{(-eps, eps)}, from a Lagrangian
interpolation of the values from the direct calculation for \eqn{\xi = }
\code{-eps}, 0 and \code{eps}.
}
\examples{
# Expected information matrices for ...
# ... sigma = 2 and xi = -0.4
gevExpInfo(2, -0.4)
# ... sigma = 3 and xi = 0.001
gevExpInfo(3, 0.001)
# ... sigma = 3 and xi = 0
gevExpInfo(3, 0)
# ... sigma = 1 and xi = 0.1
gevExpInfo(1, 0.1)

# The individual components of the latter matrix
gev11e(1, 0.1)
gev12e(1, 0.1)
gev13e(1, 0.1)
gev22e(1, 0.1)
gev23e(1, 0.1)
gev33e(0.1)
}
