% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log1pdx.R
\name{log1pdx}
\alias{log1pdx}
\alias{log1pdx2}
\title{Accurate \code{log(1+x)/x} computation (and similar)}
\usage{
log1pdx(
  x,
  minL1 = -0.79149064,
  eps2 = 0.01,
  tol_logcf = 1e-14,
  trace.lcf = FALSE,
  logCF = if (is.numeric(x)) DPQ::logcf else DPQ::logcfR.
)

log1pdx2(
  x,
  minL1 = -0.79149064,
  eps2 = 0.01,
  tol_logcf = 1e-14,
  trace.lcf = FALSE,
  logCF = if (is.numeric(x)) DPQ::logcf else DPQ::logcfR.
)
}
\arguments{
\item{x}{A numeric vector with values \eqn{x > -1}.}

\item{minL1}{A negative cutoff \eqn{m_l}. For \eqn{x \in (m_l, 1)} the
computations are based on the series expansion 4.1.29 on page 68 of
Abramowitz and Stegun (1972).}

\item{eps2}{A non-negative cutoff \eqn{\epsilon_2}. For
\eqn{|x| > \epsilon_2}, \code{\link[DPQ:logcf]{logcf}} is used in the computation
of \eqn{\log(1+x)/x} using the series expansion. Otherwise, that is,
\eqn{|x| \leq \epsilon_2}, only a few terms of the expansion are used.}

\item{tol_logcf}{A non-negative number indicating the tolerance (maximal
relative error) for the \code{\link[DPQ:logcf]{logcf}} function.}

\item{trace.lcf}{A logical scalar. Used in
\code{\link[DPQ:logcf]{logcf}}\verb{(.., trace = trace.lcf)}.}

\item{logCF}{The function to be used as \code{\link[DPQ:logcf]{logcf}}. The default
chooses the pure \R \code{logcfR()} when \code{x} is not numeric, and chooses the
C-based \code{logcf()} when \code{is.numeric(x)} is \code{TRUE}.}
}
\value{
A numeric vector (with the same attributes as \code{x}).
}
\description{
Compute \eqn{\log(1+x)/x} accurately also for small \eqn{x}, that is,
\eqn{|x| \ll 1}{|x| << 1}.  Similarly, compute
\eqn{\log(1+x)/x^2 - x^{-1}(1+x)^{-1}}.
}
\details{
For \eqn{x} in \eqn{(m_l, 1)} the computations are based on the
series expansion
\deqn{\log(1+x) = 2 \left( t+\frac{t^3}{3}+\frac{t^5}{5}+\cdots \right),}
where \eqn{t = x/(2+x)} which is formula 4.1.29 on page 68 of
Abramowitz and Stegun (1972).

For \code{log1pdx}, different computations are used in 3 different ranges of
\eqn{x}, that is,
\itemize{
\item \eqn{x < m_l} or \eqn{x > 1}: \verb{log1pdx(x):= } \code{\link{log1p}} \verb{/ x};
\item \eqn{|x| \leq \epsilon_2}:
\eqn{\frac{2}{2+x}((((\frac19y+\frac17)y+\frac15)y+\frac13)y+1)};
\item \eqn{x \in (m_l, 1)} and \eqn{|x| > \epsilon_2}:
\eqn{\frac{2}{2+x}} \code{\link{logcf}}\eqn{(y, 1, 2)}, where \eqn{y = t^2}.
}
}
\examples{
# In the limit as x tends to 0, log(1+x)/x = 1
log1pdx(0)

# log1p(x) / x is fine unless x is 0 or extremely close to 0, e.g. 1e-324,
# when it returns NaN
x <- seq(from = -1, to = 2, by = 0.01)
y1 <- log1pdx(x)
y2 <- log1p(x) / x
y <- cbind(y1, y2)
matplot(x, y, type = "l")
}
\references{
Abramowitz, M. and Stegun, I. A. (1972) Handbook of Mathematical
Functions. New York: Dover.
\url{https://en.wikipedia.org/wiki/Abramowitz_and_Stegun}
provides links to the full text, which is in public domain.
}
\author{
Paul Northrop created this function by modifying the code in
\code{\link[DPQ:log1pmx]{log1pmx}}.
}
